Create ELF aliases for identical symbols
This change greatly reduces the number of modules that need to be
compiled. The only issue right now is that sometimes when viewing
symbol table entries, the aliased symbol is chosen.

+++ libc/intrin/bswap.h
@@ -5,7 +5,7 @@ COSMOPOLITAN_C_START_
 
 uint16_t bswap_16(uint16_t) pureconst;
 uint32_t bswap_32(uint32_t) pureconst;
-uint32_t bswap_64(uint32_t) pureconst;
+uint64_t bswap_64(uint64_t) pureconst;
 
 #if defined(__GNUC__) && !defined(__STRICT_ANSI__)
 #define bswap_16(x) __builtin_bswap16(x)
